ajout adresse et data par défaut

+++ src/service/user-service.ts
@@ -1,4 +1,3 @@
-
 import bcrypt from "bcrypt-nodejs";
 import { getRepository, Repository } from "typeorm";
 
@@ -8,16 +7,16 @@ import { TokenService } from "./token-service";
 import { EmailService } from "./email-service";
 
 export class UserService {
-
-  private repository: Repository<User>
+  private repository: Repository<User>;
   constructor() {
     this.repository = getRepository(User);
   }
 
   async create(user: User): Promise<User> {
-   
-    let dbUser = await this.repository.findOne({ where: [{ email: user.email, deleted: false }] });
-    if (dbUser) throw 'Email already exist';
+    let dbUser = await this.repository.findOne({
+      where: [{ email: user.email, deleted: false }]
+    });
+    if (dbUser) throw "Email already exist";
 
     user.password = this.generatePassword(user.password);
     user.deleted = false;
@@ -38,28 +37,29 @@ export class UserService {
   }
 
   async getById(id: string): Promise<User> {
-    return await this.repository.findOneOrFail({ where: [{ id: id }], relations: ["roles"] })
+    return await this.repository.findOneOrFail({
+      where: [{ id: id }],
+      relations: ["roles"]
+    });
   }
 
-  async getByEmail(email: string) : Promise<void | User>{
-    return await this.repository.findOne({where: [{email: email}]});
+  async getByEmail(email: string): Promise<void | User> {
+    return await this.repository.findOne({ where: [{ email: email }] });
   }
 
-  async renewPassword(token: string, newPassword: string){
+  async renewPassword(token: string, newPassword: string) {
     let tokenService = new TokenService();
     let userService = new UserService();
 
     let tok = await tokenService.getLostPasswordToken(token);
-    if(!tok)
-      throw 'Token expired or not found'
-    
-    let user = await userService.getByEmail(tok.email)
-    if(!user)
-      throw 'Unable to find user'
+    if (!tok) throw "Token expired or not found";
 
-    user.password = this.generatePassword(newPassword)
+    let user = await userService.getByEmail(tok.email);
+    if (!user) throw "Unable to find user";
+
+    user.password = this.generatePassword(newPassword);
     this.repository.save(user).then(user => {
-      tokenService.deleteToken(tok)
+      tokenService.deleteToken(tok);
     });
   }
 
@@ -71,13 +71,13 @@ export class UserService {
       lastname: user.lastname,
       pseudo: user.pseudo,
       birthdate: user.birthdate
-    }
+    };
 
-    let token = jwt.sign(JSON.stringify(tokentContent), 'superspirit')
+    let token = jwt.sign(JSON.stringify(tokentContent), "superspirit");
     return token;
   }
 
-  private generatePassword(clearString: string) : string{
+  generatePassword(clearString: string): string {
     let salt = bcrypt.genSaltSync(10);
     let hash = bcrypt.hashSync(clearString, salt);
     return hash;
